Fix lint and move tests to test_deletes
